requirements.txt
Django~=5.2.3
djangorestframework>=3.15,<4.0
django-cors-headers>=4.3,<5.0
gunicorn>=21.2,<22.0
whitenoise>=6.6,<7.0
python-dotenv>=1.0,<2.0
Pillow>=10.3,<11.0
psycopg2-binary>=2.9,<3.0
dj-database-url>=2.2,<3.0
cloudinary>=1.40,<2.0
django-cloudinary-storage>=0.3,<1.0
argon2-cffi>=23.1.0